check: Add pwd as safe git directory (#5)
This is required when running the script within a container image or
prow job otherwise the following error is shown:

```
$ COMMON_INSTANCETYPES_CRI=docker make check
./scripts/cri.sh  "./scripts/check.sh"
warning: Not a git repository. Use --no-index to compare two paths outside a working tree
```

A previous attempt to fix this in the Dockerfile is removed as the
assumed path of the repo changes between local docker runs and prow
jobs.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>