Introduce hyperscale public cloud style instance types
This change introduces the following instance types modelled on those
provided by hyperscale public cloud providers:

CX
--

The CX Series provides exclusive compute resources for compute intensive
applications.

*CX* is the abbreviation of "Compute Exclusive".

The exclusive resources are given to the compute threads of the VM. In
order to ensure this, some additional cores (depending on the number of
disks and NICs) will be requested to offload the IO threading from cores
dedicated to the workload.  In addition, in this series, the NUMA
topology of the used cores is provided to the VM.

GN
--

The GN Series provides instances types intended for VMs with NVIDIA GPU
resources attached.

*GN* is the abbreviation of "GPU NVIDIA".

This series is intended to be used with VMs consuming GPUs provided by
the [NVIDIA GPU Operator](https://github.com/NVIDIA/gpu-operator) which
is made available on OpenShift via OperatorHub.

M
-

The M Series provides resources for memory intensive applications.

*M* is the abbreviation of "Memory".

N
-

The N Series is quite neutral and provides resources for general purpose
applications.

*N* is the abbreviation for "Neutral", hinting at the neutral attitude
towards workloads.

VMs of instance types will share physical CPU cores on a time-slice
basis with other VMs.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>