Drop use of dedicatedCPUPlacement
A recent change to KubeVirt [1][2] has essentially disabled the use of
dedicatedCPUPlacement as we have no way of expressing the required
resource requirements through instance types at present.

Until this is corrected all requests to create instance types with this
attribute will be rejected. This change drops the use of this attribute
until this is resolved.

[1] https://github.com/kubevirt/kubevirt/pull/8886
[2] https://github.com/kubevirt/kubevirt/issues/8867

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>