schema: Move generated schema cache out of the main tree
Carrying the generated KubeVirt JSON schema in tree is getting
cumbersome as it can change throughout the development cycle and needs
constant refreshing to be valid.

This change drops this local cache from the tree on main causing it to
be fetched before kubeconform is called as part of the validate Makefile
target. The _schemas/main/* directories being specifically added to
.gitignore to allow the cache to still be used by pinned releases of our
own release branches as these are not expected to change anywhere near
as often as main.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>