Use n1.medium during preference functional tests
Given that all preferences now have resource requirements we need to
actually provide adequate resources to the VirtualMachine.

A negative test is also added for each preference ensuring that requests
to create a VirtualMachine using an instance type that doesn't provide
enough resources are rejected.

Additionally the use of virtctl is introduced, removing a long standing
TODO from both the preference and instance type tests.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>