chore: Always add instancetype.kubevirt.io/common-instancetypes-version
This change moves to always add the version annotation to generated
resources. This is now easier as the bundles are no longer carried in
the repo and so we can add the correct commit during generation.

The release workflow is also changed to use the generate makefile target
to orchestrate the annotation of resources with the release version.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>