feat: Add preference validation tests
The functests are refactored and tests that validate preferences by
booting guests with the according OS are added. The added tests can be
skipped or focussed by the added FUNCTEST_EXTRA_ARGS variable, which
can be defined when running the cluster-functest or kubevirt-functest
targets.

E.g. run the following to skip all preference validation tests:

FUNCTEST_EXTRA_ARGS='--ginkgo.skip="VirtualMachine using a preference is able to boot"' make cluster-functest

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>