fedora: Remove interface-multiqueue
This requires >= 2 vCPUs exposed as sockets in order to be activated
within the resulting guest. At present the preference requires only a
single vCPU and doesn't provide a preferredCPUTopology.

To avoid changing already defined instance type defaults when using the
fedora preference we simply drop the interface-multiqueue tunable from
the preference for now.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>