kubevirt: Support kind kubevirtci providers
Kind providers currently expect to be called from the root of the
kubevirtci directory. This change supports them and the k8s providers by
switching into this directory before returning after each command.

The sync command is also passed the original common-instancetypes
directory and now uses the `-k` option of `kubectl apply` to build and
apply resources in a single step.

The aim is to then use this support to provide CI coverage for new arm64
based preferences in the near future.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>