image: Request a specific golang version in the builder Dockerfile
To do this we switch over to the currently Fedora 41 based golang image provided
by kubevirtci and use the GIMME_GO_VERSION ENV variable to request a
specific version of golang.

This should allow us to bump the version when required in the future
causing the post submit job to fire rebuilding the actual builder image.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>