tests: Add flag for expectedVendor to accommodate downstream modifications
Downstream vendors can obviously change the value of this label and as
the unit tests currently run against the generated bundle we need to
accommodate this fact in our unit tests.

This change adds a simple flag to allow anyone running the unit tests
downstream to provide their own expected value for the vendor label.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>